% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gks.R
\name{getGKSDataRef}
\alias{getGKSDataRef}
\title{Отображает список доступной статистики Росстата, позволяет производить навигацию по списку и получить ссылку на выбранный документ}
\usage{
getGKSDataRef()
}
\value{
Расположение файла на сайте Росстата
}
\description{
Отображает список доступной статистики Росстата, позволяет производить навигацию по списку и получить ссылку на выбранный документ
}
\examples{
getGKSDataRef()
}
