% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gks.R
\name{getTableFromDocx}
\alias{getTableFromDocx}
\title{Позволяет получить данные в виде фрейма из документа в формате docx, где есть 1 таблица с данными}
\usage{
getTableFromDocx(word_doc)
}
\arguments{
\item{word_doc}{- расположение .docx файла}
}
\value{
В переменную dataGKS записывает данные в виде фрейма
}
\description{
Позволяет получить данные в виде фрейма из документа в формате docx, где есть 1 таблица с данными
}
