% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gks.R
\name{getTableFromHtm}
\alias{getTableFromHtm}
\title{Позволяет получить данные в виде фрейма из htm/html страницы, где есть 1 таблица с данными}
\usage{
getTableFromHtm(ref)
}
\arguments{
\item{word_doc}{- ссылка на htm/html страницу, на сайте Росстата}
}
\value{
В переменную dataGKS записывает данные в виде фрейма
}
\description{
Позволяет получить данные в виде фрейма из htm/html страницы, где есть 1 таблица с данными
}
