% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gks.R
\name{info_region}
\alias{info_region}
\title{Возвращает данные в виде временного ряда по региону}
\usage{
info_region(dann, ind_name = "", id = 1)
}
\arguments{
\item{dann}{- фрейм данных, c кодами ОКАТО в последнем столбце}

\item{ind_name}{- название показателя}

\item{id}{- индекс столбца, содержащего названия регионов, по умолчанию 1}
}
\value{
Список: 
[[1]] - Фрейм, содержащий название региона и код ОКАТО
[[2]] - Название показателя
[[3]] - Фрейм, содержащий года и значения показателя по годам
}
\description{
Возвращает данные в виде временного ряда по региону
}
\examples{
info_region(Add_OKATO(dataGKS), ind_name = "Население")
}
