% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gks.R
\name{loadGKSData}
\alias{loadGKSData}
\title{Запускает функции, которые позволяют получить данные в виде фрейма, в зависимости от типа источника}
\usage{
loadGKSData(ref)
}
\arguments{
\item{ref}{- ссылка на .doc файл, на сайте Росстата, результат функции getGKSDataRef()}
}
\description{
Запускает функции, которые позволяют получить данные в виде фрейма, в зависимости от типа источника
}
\examples{
loadGKSData(getGKSDataRef())
}
