% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gks.R
\name{Return_data}
\alias{Return_data}
\title{Возвращает данные из сборника Росстата в виде временного ряда по региону или в виде панельных данных с кодами ОКАТО}
\usage{
Return_data(year = 2016)
}
\arguments{
\item{year}{- год сборника}
}
\value{
фрейм данных с кодами ОКАТО или список, результат функции info_region
}
\description{
Возвращает данные из сборника Росстата в виде временного ряда по региону или в виде панельных данных с кодами ОКАТО
}
\examples{
Return_data()
}
