% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gks.R
\docType{package}
\name{RGksStatData}
\alias{RGksStatData}
\alias{RGksStatData-package}
\title{Этот пакет позволяет импортировать данные с сайта Росстата (http://www.gks.ru).}
\description{
Этот пакет позволяет импортировать данные с сайта Росстата (http://www.gks.ru).
}
