% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gks.R
\name{toLocalEncoding}
\alias{toLocalEncoding}
\title{Исправляет кодировку для загруженного фрейма данных}
\usage{
toLocalEncoding(x, sep = ",", quote = TRUE, encoding = "utf-8")
}
\arguments{
\item{x}{- data.frame с данными
sep - разделитель полей в csv файле
quote - экранирование данных кавычками
encoding - нужная кодировка}
}
\description{
Исправляет кодировку для загруженного фрейма данных
}
\examples{
toLocalEncoding(dataGKS)
}
