% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gks.R
\name{Transform_doc_to_docx}
\alias{Transform_doc_to_docx}
\title{Позволяет получить документ в формате .docx, из документа .doc, находящегося на сайте Росстата}
\usage{
Transform_doc_to_docx(word_doc)
}
\arguments{
\item{word_doc}{- ссылка на .doc файл, на сайте Росстата}
}
\value{
Расположение .docx файла
}
\description{
Позволяет получить документ в формате .docx, из документа .doc, находящегося на сайте Росстата
}
\examples{
ref <- getGKSDataRef()
2016
3
2
ref <- gsub("/\%3Cextid\%3E/\%3Cstoragepath\%3E::\\\\|","",ref)
Transform_doc_to_docx(ref)
}
